+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Frequenz Client Base Library Release Notes
 
+## Features
+
+* Added support for HMAC signing of client messages
+
 ## Upgrading
 
 * Updated `protobuf` dependency range: changed from `>=4.21.6, <6` to `>=5.29.2, <7`